テストを通す

+++ src/main/kotlin/cndjp/qicoo/api/response/quesion/QuestionResponse.kt
@@ -7,7 +7,7 @@ class QuestionResponse(questionDTO: QuestionDTO) {
     val program_name: String
     val event_name: String
     val display_name: String
-    val like_count: Long
+    val like_count: Int
     val comment: String
     val created: String
     val updated: String
