+++ src/main/kotlin/cndjp/qicoo/domain/dto/question_aggr/QuestionDTO.kt
@@ -6,7 +6,7 @@ data class QuestionDTO(
     val program_name: String,
     val event_name: String,
     val display_name: String,
-    val like_count: Long,
+    val like_count: Int,
     val comment: String,
     val created: DateTime,
     val updated: DateTime
